@@ -12,7 +12,6 @@ using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
@@ -20,6 +19,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.IdentityModel.Tokens;
 using static Admin.API.Helpers.Extensions;
+using Microsoft.Extensions.Hosting;
 
 namespace Admin.API
 {
@@ -33,11 +33,28 @@ namespace Admin.API
 
         public IConfiguration Configuration { get; }
 
-        // This method gets called by the runtime. Use this method to add services to the container.
+        public void ConfigureDevelopmentServices(IServiceCollection services)
+        {
+
+            services.AddDbContext<AdminContext>(x => x.UseMySql(Configuration.GetConnectionString("DefaultConnection")));
+            ConfigureServices(services);
+        }
+
+        public void ConfigureProductionServices(IServiceCollection services)
+        {
+
+            services.AddDbContext<AdminContext>(x => x.UseSqlServer(Configuration.GetConnectionString("DefaultConnection")));
+            ConfigureServices(services);
+        }
+
+
+
         public void ConfigureServices(IServiceCollection services)
         {
 
             services.AddDbContext<AdminContext>(x => x.UseMySql(Configuration.GetConnectionString("DefaultConnection")));
+            // ConfigureServices(services);
+
             // // No hizo ni madres esto
             // CultureInfo.CurrentCulture = new CultureInfo("es-ES");
             IdentityBuilder builder = services.AddIdentityCore<User>(opt =>
@@ -81,22 +98,34 @@ namespace Admin.API
             options.AddPolicy("VipOnly", policy => policy.RequireRole("VIP"));
         });
 
+            services.AddControllers(options =>
+            {
+                var policy = new AuthorizationPolicyBuilder()
+    .RequireAuthenticatedUser()
+    .Build();
+                options.Filters.Add(new AuthorizeFilter(policy));
+            })
+            .AddNewtonsoftJson(options =>
+            {
 
-            services.AddMvc(options =>
-             {
-                 // Con esto pedimos autorizacion para todo el proyecto en lugar de un controlador en especifico.
+                options.SerializerSettings.ReferenceLoopHandling =
+                                    Newtonsoft.Json.ReferenceLoopHandling.Ignore;
+            });
+            // services.AddMvc(options =>
+            //  {
+            //      // Con esto pedimos autorizacion para todo el proyecto en lugar de un controlador en especifico.
 
-                 var policy = new AuthorizationPolicyBuilder()
-                  .RequireAuthenticatedUser()
-                  .Build();
-                 options.Filters.Add(new AuthorizeFilter(policy));
-             })
-             .SetCompatibilityVersion(CompatibilityVersion.Version_2_2)
-             .AddJsonOptions(opt =>
-             {
-                 opt.SerializerSettings.ReferenceLoopHandling =
-                     Newtonsoft.Json.ReferenceLoopHandling.Ignore;
-             });
+            //      var policy = new AuthorizationPolicyBuilder()
+            //       .RequireAuthenticatedUser()
+            //       .Build();
+            //      options.Filters.Add(new AuthorizeFilter(policy));
+            //  })
+            //  .SetCompatibilityVersion(CompatibilityVersion.Version_3_0)
+            //  .AddJsonOptions(opt =>
+            //  {
+            //      opt.SerializerSettings.ReferenceLoopHandling =
+            //          Newtonsoft.Json.ReferenceLoopHandling.Ignore;
+            //  });
 
             services.AddCors();
             services.Configure<CloudinarySettings>(Configuration.GetSection("CloudinarySettings"));
@@ -113,11 +142,11 @@ namespace Admin.API
             services.AddScoped<IEquipoRepository, EquipoRepository>();
             services.AddScoped<ITicketRepository, TicketRepository>();
             services.AddScoped<LogUserActivity>();
-            
+
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, Seed seeder, ILogger<Startup> logger)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, Seed seeder, ILogger<Startup> logger)
         {
             if (env.IsDevelopment())
             {
@@ -144,20 +173,22 @@ namespace Admin.API
             }
             // El orden es importante
             // app.UseHttpsRedirection();
-             seeder.SeedUsers();
+            app.UseRouting();
+            app.UseAuthentication();
+            app.UseAuthorization();
+            seeder.SeedUsers();
             // Hay que cambiar esto ya que permite TODO
             app.UseCors(x => x.AllowAnyOrigin().AllowAnyHeader().AllowAnyMethod());
-            app.UseAuthentication();
+
             app.UseDefaultFiles();
             app.UseStaticFiles();
-            app.UseMvc(routes =>
-            {
 
-                routes.MapSpaFallbackRoute(
-                    name: "spa-fallback",
-                    defaults: new { controller = "Fallback", action = "Index" }
-                );
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+                endpoints.MapFallbackToController("Index", "Fallback");
             });
+
         }
     }
 }
